requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=2.0.0
markdown2>=2.4.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
python-slugify>=8.0.0
